+++ packages/wallet/src/initialization/instances/index.ts
@@ -1,4 +1,5 @@
 export { accountsController } from './accounts-controller';
+export { approvalController } from './approval-controller/approval-controller';
 export { connectivityController } from './connectivity-controller';
 export { keyringController } from './keyring-controller';
 export { storageService } from './storage-service';