+++ packages/wallet/src/types.ts
@@ -8,6 +8,7 @@ import type {
   DefaultEvents,
   RootMessenger,
 } from './initialization/defaults';
+import type { ApprovalControllerInstanceOptions } from './initialization/instances/approval-controller/types';
 import { GenericEncryptor } from './initialization/instances/keyring-controller';
 import { InitializationConfiguration } from './initialization/types';
 
@@ -22,6 +23,7 @@ export type WalletOptions = {
 };
 
 export type InstanceSpecificOptions = {
+  approvalController?: ApprovalControllerInstanceOptions;
   connectivityController: {
     connectivityAdapter: ConnectivityAdapter;
   };